Module for Manager
Version of the module and the project is updated to v.0.1.2
