+++ Manager/ModuleMAN1/src/main/java/org/ModuleMAN1/Domain/Control_Assignment.java
@@ -1,4 +1,4 @@
-package A02.Manager.Domain;
+package org.ModuleMAN1.Domain;
 
 /** 
  * <p>
