+++ Manager/ModuleMAN1/src/main/java/org/ModuleMAN1/Domain/Member_Types.java
@@ -1,4 +1,4 @@
-package A02.Manager.Domain;
+package org.ModuleMAN1.Domain;
 
 public enum Member_Types {
 	_teacher, _student;
