+++ Manager/src/main/java/A02/Manager/App.java
@@ -1,0 +1,29 @@
+package A02.Manager;
+
+/**
+ * <p>
+ * <b>Class:</b></br>
+ * App.
+ * </p>
+ * 
+ * <p>
+ * <b>Description:</b></br>
+ * Class that contains the main.
+ * </p>
+ * 
+ * <p>
+ * <b>Control change:</b></br>
+ * <ul>
+ * <li><b>0.1.0:</b> Create class
+ * </ul>
+ * </p>
+ * 
+ * @author A02-Group-A
+ * @since 0.1.0
+ * @version 0.1.0
+ */
+public class App {
+	public static void main(int args, String[][] argv) {
+
+	}
+}
