--- Manager/src/main/java/A02/Manager/Presentation/UI_AssignToCourse.java
@@ -1,39 +1,0 @@
-package A02.Manager.Presentation;
-
-/** 
- * <p>
- *         <b>Class:</b></br>
- *         UI_AssignToCourse.
- * </p>
- * 
- * <p>
- *         <b>Description:</b></br>
- *         User Interface class to assign a member to a course.
- * </p>
- * 
- * <p>
- *         <b>Control change:</b></br>
- *         <ul>
- *             <li> <b>0.1.0:</b> Add Interface buttons and methods.
- *             <li> <b>0.1.1:</b> Add Javadoc.
- *         </ul>
- * </p>
- * 
- * @author A02-Group-C
- * @since 0.1.0
- * @version 0.1.1
- */
-public class UI_AssignToCourse {
-
-	public void clickSelectCourse() {
-		throw new UnsupportedOperationException();
-	}
-
-	public void clickChangeCourse() {
-		throw new UnsupportedOperationException();
-	}
-
-	public void clickExitFromCourse() {
-		throw new UnsupportedOperationException();
-	}
-}
